Add GitHub Action to generate offline dataset
This Action simply tars up the JSON files generated by `./_plugins/create-json-files.rb` and uploads them to a release called "latest", replacing it if it already exists.

This provides a simple offline dataset for programmatic use, without requiring one to query the public API and without using up its quota.

It runs on every push to the `main` and `master` branches, and can also be triggered manually.